+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2024, MarcGebhardt'
 author = 'MarcGebhardt'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.0'
+release = 'v0.1.1'
 
 
 # -- General configuration ---------------------------------------------------
@@ -62,6 +62,7 @@ napoleon_attr_annotations = True
 # a list of builtin themes.
 #
 html_theme = 'sphinx_rtd_theme'
+html_theme_options = {'navigation_depth': 6}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,