+++ CHANGELOG.md
@@ -7,13 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [v0.1.1] – 2023-01-25
+
 ### Added
 
 - Packing of log-files of series (common.loadnsave.comb_logs)
 - Identifying type of given object (standard, numpy and pandas; common.helper.type_str_return)
 - Relative deviation implemented (common.stat_ext.relative_deviation)
 - Coefficient of variation (with or without statistical outliers) with short named function (common.stat_ext.cv and common.stat_ext.cvwoso) for use in pandas aggregate
-- Functionality for automatic determination of points of interest added to ./common/mc_char (PoI_relFinder, PoI_PLH, POI_VIP_namer, PoI_fixeva and PoI_refinement)
+- Functionality for automatic determination of points of interest added to ./common/mc_char (poi_rel_finder, poi_det_plh, poi_vip_namer, poi_fixeva and poi_refinement)
 
 ### Changed
 
@@ -30,7 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Evaluation of three-point-bending tests of cortical bone specimens with different moisture states (./scripts/02_MMCB/, old structure), Including different improvements:
 	- improvements of measured curve characterization (MCurve_Characterizer, MCurve_Char_Plotter,...)
 	- automatic setting of load direction by optical measured traverse displacement (ldoption='auto-Pcoorddisp' in bending/fitting/Perform_Fit)
-	- automatic determination and naming of points of interest (see ./common/mc_char (PoI_relFinder, PoI_PLH, POI_VIP_namer, PoI_fixeva and PoI_refinement))
+	- automatic determination and naming of points of interest (see ./common/mc_char (poi_rel_finder, poi_det_plh, poi_vip_namer, poi_fixeva and poi_refinement)
 - Mandibulae evaluation (./scripts/03_Zesbo_Mand/, old structure, no relevant changes to other projects)
 
 ## [v0.1.0] – 2023-01-17
@@ -41,4 +44,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [unreleased]: https://github.com/MarcGebhardt/ExMechEva/tree/main
+[v0.1.1]: https://github.com/MarcGebhardt/ExMechEva/releases/tag/v0.1.1
 [v0.1.0]: https://github.com/MarcGebhardt/ExMechEva/releases/tag/v0.1.0